Moved directive to its own file with a register wrapper so we can pass in angular just once.
